Move to foundry devops for keeping track of latest addresses (#754)
This PR updates our makefile deployment scripts to be more modular and
easier to understand. Plays huge part in removal and addition of facets
in upcoming PR for alpha deployment.

---------

+++ contracts/src/utils/Validator.sol
@@ -16,7 +16,7 @@ library Validator {
     if (byteName.length == 0) revert Validator__InvalidStringLength();
   }
 
-  function checkLength(string memory name, uint min) internal pure {
+  function checkLength(string memory name, uint256 min) internal pure {
     bytes memory byteName = bytes(name);
     if (byteName.length < min) revert Validator__InvalidStringLength();
   }
